ScottPlot.Ticks namespace
