+++ src/ScottPlot.Avalonia/AvaPlot.axaml.cs
@@ -16,7 +16,7 @@ using Avalonia.Media;
 using Avalonia.Platform;
 using Avalonia.VisualTree;
 using ScottPlot.Avalonia;
-using ScottPlot.Config;
+using ScottPlot.Ticks;
 using ScottPlot.Interactive;
 using Ava = Avalonia;
