+++ src/ScottPlot.Demo/Customize/Ticks.cs
@@ -329,7 +329,7 @@ namespace ScottPlot.Demo.Customize
                 plt.Ticks(dateTimeX: true);
 
                 // define tick spacing as 1 day (every day will be shown)
-                plt.Grid(xSpacing: 1, xSpacingDateTimeUnit: Config.DateTimeUnit.Day);
+                plt.Grid(xSpacing: 1, xSpacingDateTimeUnit: ScottPlot.Ticks.DateTimeUnit.Day);
                 plt.Ticks(dateTimeX: true, xTickRotation: 45);
                 plt.Layout(bottom: 60);
             }
