+++ src/ScottPlot/Plottable/Annotation.cs
@@ -1,4 +1,4 @@
-﻿using ScottPlot.Config;
+﻿using ScottPlot.Ticks;
 using ScottPlot.Drawing;
 using ScottPlot.Renderable;
 using System;
