+++ src/ScottPlot/Plottable/FinancePlot.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Drawing;
-using ScottPlot.Config;
+using ScottPlot.Ticks;
 using ScottPlot.Drawing;
 using ScottPlot.Renderable;
