+++ src/ScottPlot/Plottable/Image.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel;
 using System.Drawing;
-using ScottPlot.Config;
+using ScottPlot.Ticks;
 using ScottPlot.Drawing;
 using ScottPlot.Renderable;
