+++ src/ScottPlot/Plottable/RadarPlot.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
-using ScottPlot.Config;
+using ScottPlot.Ticks;
 using ScottPlot.Drawing;
 using ScottPlot.Renderable;
