+++ src/ScottPlot/Ticks/DateTimeTickUnits/MonthToMinutes/DateTimeTickHour.cs
@@ -1,25 +1,25 @@
 ﻿using System;
 using System.Globalization;
 
-namespace ScottPlot.Config.DateTimeTickUnits
+namespace ScottPlot.Ticks.DateTimeTickUnits
 {
-    public class DateTimeTickMinute : DateTimeTickUnitBase
+    public class DateTimeTickHour : DateTimeTickUnitBase
     {
-        public DateTimeTickMinute(CultureInfo culture, int maxTickCount, int? manualSpacing) : base(culture, maxTickCount, manualSpacing)
+        public DateTimeTickHour(CultureInfo culture, int maxTickCount, int? manualSpacing) : base(culture, maxTickCount, manualSpacing)
         {
-            kind = DateTimeUnit.Minute;
+            kind = DateTimeUnit.Hour;
             if (manualSpacing == null)
-                deltas = new int[] { 1, 2, 5, 10, 15, 30 };
+                deltas = new int[] { 1, 2, 4, 8, 12 };
         }
 
         protected override DateTime Floor(DateTime value)
         {
-            return new DateTime(value.Year, value.Month, value.Day, value.Hour, 0, 0);
+            return new DateTime(value.Year, value.Month, value.Day);
         }
 
         protected override DateTime Increment(DateTime value, int delta)
         {
-            return value.AddMinutes(delta);
+            return value.AddHours(delta);
         }
 
         protected override string GetTickLabel(DateTime value)
