+++ src/ScottPlot/Ticks/DateTimeTickUnits/MonthToMinutes/DateTimeTickMonth.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Globalization;
 
-namespace ScottPlot.Config.DateTimeTickUnits
+namespace ScottPlot.Ticks.DateTimeTickUnits
 {
     public class DateTimeTickMonth : DateTimeTickUnitBase
     {
