+++ src/ScottPlot/Ticks/DateTimeUnit.cs
@@ -1,4 +1,4 @@
-﻿namespace ScottPlot.Config
+﻿namespace ScottPlot.Ticks
 {
     public enum DateTimeUnit { ThousandYear, HundredYear, TenYear, Year, Month, Day, Hour, Minute, Second, Decisecond, Centisecond, Millisecond };
 }
